Reorganise the `version` module
